feat(beacon-node): run consensus-specs fork-choice compliance tests
Adds a new `compliance_fork_choice` spec test runner alongside the
existing `fork_choice` runner. Tests are downloaded out-of-band from
the consensus-specs "Compliance Tests" workflow artifact and exercised
via the same fork-choice driver, with two test-only accommodations for
fixture conventions:

- Honor `meta.bls_setting == 2` (placeholder signatures) by passing
  `validSignatures: true` to `chain.processBlock`.
- Treat `BLOCK_ERROR_ALREADY_KNOWN` as a no-op success when a step
  expects validity, matching spec semantics for `on_block(store, block)`
  on a duplicate (compliance fixtures intentionally re-import blocks
  via `dup_shift` mutations).

The existing `forkChoiceTest` factory was extracted from
`fork_choice.test.ts` into `test/spec/utils/forkChoiceRunner.ts` so
both runners share it; the original test file is now a thin entrypoint.

Tooling:
- `scripts/download-compliance-fc-tests.sh` resolves the artifact via
  `--dir`, `--tarball`, `--url`, `--run-id`, or auto-fetches the latest
  successful workflow run via `gh`. Mirrors Prysm's
  `hack/compliance-fc-report.sh` pattern. Magic-byte sniffing handles
  both zip-wrapped and raw `tar.gz` artifacts (the consensus-specs
  workflow currently publishes the latter).
- `scripts/compliance-fc-report.sh` runs the suite under vitest with
  `--reporter=json` and prints a per-suite total/pass/fail/skip table
  plus an error-grouped failure breakdown.

Current pass rate against `small.tar.gz`: 506 / 2944 (~17%), ahead of
the comparable Prysm baseline (~12%). Both `block_cover` suites pass
fully (192/192 each fork). Remaining failure classes are tracked as
follow-ups:

- 1280x SSZ deserialize mismatch on every gloas suite except
  block_cover ("First offset must equal to fixedEnd 80 != 48"). The
  artifact was generated from consensus-specs master while Lodestar's
  spec-tests pin is `v1.7.0-alpha.5`; a Gloas container shape on
  master is ahead of our `@lodestar/types` definitions. Fix belongs
  in a separate PR updating types.
- ~600x `Invalid proposer boost root at step N` — real fork-choice
  compliance gaps in fulu suites.
- 145x `EPOCH_CONTEXT_ERROR_COMMITTEE_EPOCH_OUT_OF_RANGE`.
- ~80x `FORKCHOICE_ERROR_INVALID_ATTESTATION`.

CI is intentionally not wired; the runner cleanly skips when the data
directory is absent so the existing `test:spec:*` jobs are unaffected.

AI Assistance Disclosure: Used Claude Code.
