+++ packages/beacon-node/test/spec/specTestVersioning.ts
@@ -13,3 +13,14 @@ export const blsSpecTests = {
   ...specTestVersions.blsSpecTests,
   outputDir: path.join(__dirname, "../../", specTestVersions.blsSpecTests.outputDirBase),
 };
+
+// Compliance fork choice tests are generated by a separate consensus-specs workflow
+// ("Compliance Tests") and published only as a transient GitHub Actions artifact.
+// They are downloaded out-of-band via `pnpm download-compliance-fc-tests`.
+// Generator configs: `tiny`, `small`, `standard`. The scheduled workflow currently
+// publishes only `small.tar.gz`.
+export const complianceForkChoiceSpecTests = {
+  outputDirBase: "spec-tests-compliance",
+  outputDir: path.join(__dirname, "../../", "spec-tests-compliance"),
+  configs: ["tiny", "small", "standard"] as const,
+};
