+++ scripts/download-compliance-fc-tests.sh
@@ -1,0 +1,233 @@
+#!/usr/bin/env bash
+#
+# Download and extract consensus-specs fork-choice compliance test data.
+#
+# These tests are NOT part of the standard `download-spec-tests` bundle; they are
+# generated by a separate scheduled workflow ("Compliance Tests") on
+# ethereum/consensus-specs and published only as a transient GitHub Actions
+# artifact (e.g. `small.tar.gz`, ~80MB). This script mirrors Prysm's
+# hack/compliance-fc-report.sh resolution order to support multiple sources.
+#
+# Resolution order (first matching wins):
+#   --dir <path>       Use a pre-extracted directory (no download)
+#   --tarball <path>   Use a local tarball
+#   --url <url>        Download a tarball from a URL
+#   --run-id <id>      Download artifact from a specific GitHub Actions run
+#   (default)          Auto-fetch latest successful "Compliance Tests" run via gh
+#
+# Usage:
+#   ./scripts/download-compliance-fc-tests.sh                  # auto-fetch small
+#   ./scripts/download-compliance-fc-tests.sh --config small
+#   ./scripts/download-compliance-fc-tests.sh --tarball ~/Downloads/small.tar.gz
+#   ./scripts/download-compliance-fc-tests.sh --run-id 24754336017
+#   ./scripts/download-compliance-fc-tests.sh --url https://example.org/small.tar.gz
+#
+# Output:
+#   packages/beacon-node/spec-tests-compliance/<config>/
+
+set -euo pipefail
+
+SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+REPO_ROOT="$(cd "$SCRIPT_DIR/.." && pwd)"
+PACKAGE_DIR="$REPO_ROOT/packages/beacon-node"
+OUTPUT_BASE="$PACKAGE_DIR/spec-tests-compliance"
+CACHE_DIR="${LODESTAR_COMPLIANCE_FC_CACHE:-/var/tmp/lodestar_compliance_fc_cache}"
+
+CONFIG="small"
+TARBALL=""
+URL=""
+RUN_ID=""
+DIR=""
+GH_REPO="ethereum/consensus-specs"
+GH_WORKFLOW="Compliance Tests"
+
+usage() {
+  cat <<'EOF'
+Usage: download-compliance-fc-tests.sh [options]
+
+Options:
+  --config <tiny|small|standard>  Generator config to fetch (default: small)
+  --tarball <path>                Use a local .tar.gz file
+  --url <url>                     Download a .tar.gz from a URL
+  --run-id <id>                   Download artifact from a specific GH Actions run
+  --dir <path>                    Use a pre-extracted directory
+  --repo <owner/name>             Override consensus-specs repo (default: ethereum/consensus-specs)
+  --workflow <name>               Override workflow name (default: "Compliance Tests")
+  -h, --help                      Show this help
+
+Environment:
+  LODESTAR_COMPLIANCE_FC_CACHE    Tarball cache dir (default: /var/tmp/lodestar_compliance_fc_cache)
+  GITHUB_TOKEN                    Required by `gh` for non-public artifact downloads
+EOF
+}
+
+while [[ $# -gt 0 ]]; do
+  case "$1" in
+    --config)   CONFIG="$2"; shift 2 ;;
+    --tarball)  TARBALL="$2"; shift 2 ;;
+    --url)      URL="$2"; shift 2 ;;
+    --run-id)   RUN_ID="$2"; shift 2 ;;
+    --dir)      DIR="$2"; shift 2 ;;
+    --repo)     GH_REPO="$2"; shift 2 ;;
+    --workflow) GH_WORKFLOW="$2"; shift 2 ;;
+    -h|--help)  usage; exit 0 ;;
+    *)          echo "Unknown argument: $1" >&2; usage; exit 1 ;;
+  esac
+done
+
+case "$CONFIG" in
+  tiny|small|standard) ;;
+  *) echo "Invalid --config: $CONFIG (must be tiny, small, or standard)" >&2; exit 1 ;;
+esac
+
+OUTPUT_DIR="$OUTPUT_BASE/$CONFIG"
+ARTIFACT_NAME="${CONFIG}.tar.gz"
+
+require_gh() {
+  command -v gh >/dev/null 2>&1 || {
+    echo "error: 'gh' CLI is required for this resolution mode" >&2
+    echo "       install from https://cli.github.com or provide --tarball/--url/--dir instead" >&2
+    exit 1
+  }
+}
+
+extract_to_output() {
+  local tarball="$1"
+  if [[ ! -f "$tarball" ]]; then
+    echo "error: tarball not found at $tarball" >&2
+    exit 1
+  fi
+  echo "Extracting $tarball -> $OUTPUT_DIR"
+  rm -rf "$OUTPUT_DIR"
+  mkdir -p "$OUTPUT_DIR"
+  tar -xzf "$tarball" -C "$OUTPUT_DIR"
+}
+
+copy_dir_to_output() {
+  local src="$1"
+  if [[ ! -d "$src" ]]; then
+    echo "error: --dir path not found at $src" >&2
+    exit 1
+  fi
+  echo "Copying $src -> $OUTPUT_DIR"
+  rm -rf "$OUTPUT_DIR"
+  mkdir -p "$OUTPUT_DIR"
+  cp -R "$src"/. "$OUTPUT_DIR"/
+}
+
+download_url() {
+  local url="$1"
+  local key
+  key=$(printf '%s' "$url" | shasum | awk '{print $1}')
+  local cache_path="$CACHE_DIR/url-$key"
+  mkdir -p "$cache_path"
+  local target="$cache_path/$ARTIFACT_NAME"
+  if [[ ! -f "$target" ]]; then
+    echo "Downloading $url" >&2
+    curl -fsSL "$url" -o "$target"
+  else
+    echo "Using cached $target" >&2
+  fi
+  printf '%s' "$target"
+}
+
+download_run_artifact() {
+  local run_id="$1"
+  require_gh
+  local cache_path="$CACHE_DIR/run-$run_id"
+  mkdir -p "$cache_path"
+  local target="$cache_path/$ARTIFACT_NAME"
+
+  if [[ -f "$target" ]]; then
+    echo "Using cached $target" >&2
+    printf '%s' "$target"
+    return
+  fi
+
+  # Don't use `gh run download` because it assumes artifacts are zip-wrapped
+  # by GitHub. The consensus-specs Compliance Tests workflow publishes the
+  # `*.tar.gz` blob directly, so gh's extraction step fails with
+  # "zip: not a valid zip file". Fetch via the artifacts API and sniff the
+  # magic bytes ourselves to handle either case.
+  echo "Looking up artifact $ARTIFACT_NAME in run $run_id ($GH_REPO)" >&2
+  local artifact_id
+  artifact_id=$(gh api "repos/$GH_REPO/actions/runs/$run_id/artifacts" --paginate \
+    --jq ".artifacts[] | select(.name == \"$ARTIFACT_NAME\") | .id" | head -n 1)
+
+  if [[ -z "$artifact_id" ]]; then
+    echo "error: artifact named '$ARTIFACT_NAME' not found in run $run_id" >&2
+    exit 1
+  fi
+
+  local raw="$cache_path/raw.bin"
+  echo "Downloading artifact id=$artifact_id" >&2
+  gh api "repos/$GH_REPO/actions/artifacts/$artifact_id/zip" > "$raw"
+
+  # Sniff first 2 bytes to distinguish zip (50 4b = "PK") from gzip (1f 8b).
+  local magic
+  magic=$(head -c 2 "$raw" | od -An -tx1 | tr -d ' \n')
+
+  case "$magic" in
+    504b)
+      echo "Artifact is zip-wrapped — unwrapping" >&2
+      command -v unzip >/dev/null || { echo "error: unzip required to unwrap artifact" >&2; exit 1; }
+      (cd "$cache_path" && unzip -o -q "$raw")
+      rm -f "$raw"
+      if [[ ! -f "$target" ]]; then
+        local found
+        found=$(find "$cache_path" -name '*.tar.gz' -type f -print -quit || true)
+        if [[ -z "$found" ]]; then
+          echo "error: no .tar.gz found inside zip-wrapped artifact at $cache_path" >&2
+          ls -la "$cache_path" >&2 || true
+          exit 1
+        fi
+        mv "$found" "$target"
+      fi
+      ;;
+    1f8b)
+      echo "Artifact is a raw gzip — using directly" >&2
+      mv "$raw" "$target"
+      ;;
+    *)
+      echo "error: unknown artifact format (magic=$magic) at $raw" >&2
+      head -c 200 "$raw" >&2 || true
+      exit 1
+      ;;
+  esac
+
+  printf '%s' "$target"
+}
+
+resolve_latest_run() {
+  require_gh
+  echo "Looking up latest successful '$GH_WORKFLOW' run on $GH_REPO" >&2
+  gh -R "$GH_REPO" run list \
+    --workflow "$GH_WORKFLOW" \
+    --status success \
+    --limit 1 \
+    --json databaseId \
+    --jq '.[0].databaseId'
+}
+
+if [[ -n "$DIR" ]]; then
+  copy_dir_to_output "$DIR"
+elif [[ -n "$TARBALL" ]]; then
+  extract_to_output "$TARBALL"
+elif [[ -n "$URL" ]]; then
+  TARBALL_PATH=$(download_url "$URL")
+  extract_to_output "$TARBALL_PATH"
+elif [[ -n "$RUN_ID" ]]; then
+  TARBALL_PATH=$(download_run_artifact "$RUN_ID")
+  extract_to_output "$TARBALL_PATH"
+else
+  RUN_ID=$(resolve_latest_run)
+  if [[ -z "$RUN_ID" ]]; then
+    echo "error: no successful runs of '$GH_WORKFLOW' found on $GH_REPO" >&2
+    exit 1
+  fi
+  echo "Latest successful run: $RUN_ID"
+  TARBALL_PATH=$(download_run_artifact "$RUN_ID")
+  extract_to_output "$TARBALL_PATH"
+fi
+
+echo "Done. Tests extracted under $OUTPUT_DIR"
