+++ vitest.config.ts
@@ -54,6 +54,7 @@ export default defineConfig({
     exclude: [
       "**/spec-tests/**",
       "**/spec-tests-bls/**",
+      "**/spec-tests-compliance/**",
       "**/*.browser.test.ts",
       "**/node_modules/**",
       "**/dist/**",
@@ -93,6 +94,7 @@ export default defineConfig({
         "**/node_modules/**",
         "**/spec-tests/**",
         "**/spec-tests-bls/**",
+        "**/spec-tests-compliance/**",
       ],
     },
   },
